move callbacks to separate classes
